feat(intl): Supports `Intl.supportedValuesOf`
